fix issues

+++ private_rest_res_account.go
@@ -6,7 +6,7 @@ type PrivateRestAccountDetailRes struct {
 	UserId        int64    `json:"user_id"`
 	Tier          int64    `json:"tier"`
 	Key           struct {
-		Mode int `json:"mode"` // 模式： 1 - 经典模式 2 - 旧版统一模式
+		Mode int64 `json:"mode"` // 模式： 1 - 经典模式 2 - 旧版统一模式
 	} `json:"key"`
-	CopyTradingRole int `json:"copy_trading_role"`
+	CopyTradingRole int64 `json:"copy_trading_role"`
 }
